Solution Upload
I made the changes to get the messenger system working. This should accomplish what was asked for in the README doc. Most of the changes went into the messages.js like it said, but there was a slight change to footer.js adding the value to the input in order to clear out the text when the user sent their message.